+++ packages/obj-diff/__tests__/utils.ts
@@ -2,8 +2,10 @@ import type {
   deserialize as deserializeType,
   diff as diffType,
   diffWith as diffWithType,
+  parse as parseType,
   patch as patchType,
   serialize as serializeType,
+  stringify as stringifyType,
 } from "../src/index";
 
 let diff: typeof diffType;
@@ -11,6 +13,8 @@ let diffWith: typeof diffWithType;
 let patch: typeof patchType;
 let serialize: typeof serializeType;
 let deserialize: typeof deserializeType;
+let stringify: typeof stringifyType;
+let parse: typeof parseType;
 
 if (process.env.TEST_CJS) {
   const dist = require("../dist/index.cjs");
@@ -19,6 +23,8 @@ if (process.env.TEST_CJS) {
   patch = dist.patch;
   serialize = dist.serialize;
   deserialize = dist.deserialize;
+  stringify = dist.stringify;
+  parse = dist.parse;
 } else if (process.env.TEST_DIST) {
   const dist = await import("../dist/index.js");
   diff = dist.diff;
@@ -26,6 +32,8 @@ if (process.env.TEST_CJS) {
   patch = dist.patch;
   serialize = dist.serialize;
   deserialize = dist.deserialize;
+  stringify = dist.stringify;
+  parse = dist.parse;
 } else {
   const src = await import("../src/index");
   diff = src.diff;
@@ -33,6 +41,8 @@ if (process.env.TEST_CJS) {
   patch = src.patch;
   serialize = src.serialize;
   deserialize = src.deserialize;
+  stringify = src.stringify;
+  parse = src.parse;
 }
 
-export { diff, diffWith, patch, serialize, deserialize };
+export { diff, diffWith, patch, serialize, deserialize, stringify, parse };